requirements.txt
Jinja2==2.9.6
MarkupSafe==1.0
py==1.4.34
pyaml==16.12.2
pytest==3.1.3
PyYAML==3.12